==> S2 - Python script to select best combinations from BT (Recent data).ipynb <==
import pandas as pd
import numpy as np
import os
import datetime as dt

path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/BT results v1/"

trade_data_path = "E:/WSG Markets/Backtesting V2 Nov 2025/BT results/"

folders = os.listdir(path)
# folders.remove("Seykota")
# folders.remove("Long Term Pull Back")

# folders = ['Seykota']
len(folders)

years_list = ['2025']

Major_pairs_list = [
    "AUD/CAD", "AUD/CHF", "AUD/JPY", "AUD/NZD", "AUD/SGD", "AUD/USD",
    "CAD/CHF", "CAD/JPY", "CAD/SGD",
    "CHF/JPY", "CHF/ZAR",
    "EUR/AUD", "EUR/CAD", "EUR/CHF", "EUR/GBP", "EUR/HUF", "EUR/JPY",
    "EUR/NOK", "EUR/NZD", "EUR/PLN", "EUR/SEK", "EUR/TRY", "EUR/USD", "EUR/ZAR",
    "GBP/AUD", "GBP/CAD", "GBP/CHF", "GBP/JPY", "GBP/NZD", "GBP/PLN", "GBP/USD",
    "HKD/JPY",
    "NZD/CAD", "NZD/CHF", "NZD/JPY", "NZD/USD",
    "TRY/JPY",
    "USD/CAD", "USD/CHF", "USD/CNH", "USD/DKK", "USD/JPY", "USD/MXN",
    "USD/NOK", "USD/SEK", "USD/SGD", "USD/TRY", "USD/ZAR", "ZAR/JPY"]

Exotic_pairs_list = ['EUR/NOK','EUR/SEK','USD/MXN','USD/ZAR','USD/CNH','USD/NOK','ZAR/JPY','USD/SEK']

Commodities_pairs_list = [ "XBR/USD","XTI/USD","XNG/USD","XCU/USD",'XAU/USD','XAG/USD',
                         'XAG/EUR','XAU/AUD','XAU/EUR','XAU/GBP','XAU/JPY']

ETF_pairs_list = ['AUS/USD', 'CHN/USD', 'ESP/USD', 'EUS/USD', 'FRA/USD', 'GER/USD',
                   'HKG/USD', 'JPN/USD', 'NAS/USD', 'SPX/USD', 'UK1/USD', 'US3/USD']

bond = ['FEI/USD']

futures = ['VZY/USD']

Index_CFDs = ['DXY/USD']

NDFs = ['USD/INR','USD/TWD']

Single_Stock_CFDs = ['AAY/USD','AAP/USD','ADB/USD','AMA/USD','AMG/USD',
                    'AMZ/USD','AXP/USD','BAY/USD','BAB/USD','BAC/USD','CAT/USD','CHR/USD',
                     'CMC/USD','CME/USD','COS/USD','CSC/USD','CTA/USD','CTS/USD','CVX/USD',
                     'DDY/USD','DOW/USD','EBA/USD','EXP/USD','GEN/USD','GOO/USD','GGL/USD',
                     'HDY/USD','HPQ/USD','IBM/USD','INT/USD','INU/USD','ISR/USD','JNJ/USD',
                     'JPM/USD','KHC/USD','KLA/USD','KOY/USD','LRC/USD','MCD/USD','MCH/USD',
                     'MMM/USD','MRK/USD','MSF/USD','NTA/USD','NTR/USD','PAY/USD','PCA/USD',
                     'PGY/USD','QCO/USD','ROS/USD','RTX/USD','SBU/USD','VRS/USD','WMT/USD',
                    "AAP/USD", "AAY/USD", "ABT/USD", "ADB/USD", "ADM/USD", "ADS/USD",
                    "AEP/USD", "AES/USD", "AFL/USD", "AIG/USD", "AIZ/USD", "AKA/USD",
                    "ALL/USD", "AMA/USD", "AMG/USD", "AMP/USD", "AMZ/USD", "ANF/USD",
                    "ANY/USD", "APA/USD", "ASM/USD", "AUP/USD", "AVG/USD", "AXP/USD",
                    "AZO/USD", "BAB/USD", "BAC/USD", "BAX/USD", "BAY/USD", "BBY/USD",
                    "BEN/USD", "BKY/USD", "BLK/USD", "BMO/USD", "BMY/USD", "BNS/USD",
                    "BSX/USD", "BUD/USD", "BXP/USD", "CAT/USD", "CBY/USD", "CCL/USD",
                    "CHR/USD", "CIE/USD", "CIY/USD", "CMA/USD", "CMC/USD", "CME/USD",
                    "CMY/USD", "CNP/USD", "CNQ/USD", "CNX/USD", "COF/USD", "COP/USD",
                    "COS/USD", "CPB/USD", "CRM/USD", "CSC/USD", "CSX/USD", "CTA/USD",
                    "CTS/USD", "CVE/USD", "CVS/USD", "CVX/USD", "CYY/USD", "DAL/USD",
                    "DEY/USD", "DFS/USD", "DGX/USD", "DHI/USD", "DIS/USD", "DPZ/USD",
                    "DTE/USD", "DUK/USD", "DVN/USD", "DYY/USD", "EAY/USD", "EBA/USD",
                    "EMR/USD", "ENB/USD", "EOG/USD", "EXC/USD", "EXE/USD", "EXP/USD",
                    "FAS/USD", "FCX/USD", "FDX/USD", "FEY/USD", "FIT/USD", "FOA/USD",
                    "FOX/USD", "FYY/USD", "GDY/USD", "GEY/USD", "GGO/USD", "GIL/USD",
                    "GIS/USD", "GLW/USD", "GMY/USD", "GNW/USD", "GOO/USD", "GRM/USD",
                    "GSY/USD", "HAL/USD", "HDY/USD", "HOG/USD", "HON/USD", "HPQ/USD",
                    "HSY/USD", "HUM/USD", "IBM/USD", "ICE/USD", "INT/USD", "INU/USD",
                    "IPY/USD", "ISR/USD", "JCI/USD", "JNJ/USD", "JNP/USD", "JPM/USD",
                    "KEY/USD", "KGC/USD", "KHC/USD", "KLA/USD", "KMB/USD", "KOY/USD",
                    "KSS/USD", "LEN/USD", "LLY/USD", "LMT/USD", "LOW/USD", "LRC/USD",
                    "LVS/USD", "LYY/USD", "MAR/USD", "MAT/USD", "MAY/USD", "MBI/USD",
                    "MCD/USD", "MCH/USD", "MCO/USD", "MDT/USD", "MET/USD", "MGM/USD",
                    "MMC/USD", "MMM/USD", "MOY/USD", "MRK/USD", "MRO/USD", "MSF/USD",
                    "MSI/USD", "MSY/USD", "MUY/USD", "MYY/USD", "NEE/USD", "NEM/USD",
                    "NFL/USD", "NKE/USD", "NOC/USD", "NSC/USD", "NTA/USD", "NTE/USD",
                    "NTR/USD", "NUE/USD", "NVD/USD", "ODP/USD", "ORC/USD", "OXY/USD",
                    "PAY/USD", "PBR/USD", "PCA/USD", "PDC/USD", "PEG/USD", "PEP/USD",
                    "PFE/USD", "PGR/USD", "PGY/USD", "PMY/USD", "PNC/USD", "PRU/USD",
                    "QCO/USD", "RHI/USD", "RIG/USD", "RLY/USD", "ROK/USD", "ROS/USD",
                    "SAP/USD", "SBU/USD", "SCH/USD", "SIR/USD", "SLB/USD", "SNY/USD",
                    "SOY/USD", "SPG/USD", "STT/USD", "STX/USD", "SWK/USD", "SYY/USD",
                    "TGT/USD", "TMU/USD", "TRV/USD", "TSL/USD", "TXN/USD", "TYY/USD",
                    "UAL/USD", "UAY/USD", "UBS/USD", "ULT/USD", "UNH/USD", "UNP/USD",
                    "UPS/USD", "USB/USD", "VLO/USD", "VOD/USD", "VRS/USD", "VYY/USD",
                    "VZY/USD", "WBA/USD", "WFC/USD", "WHR/USD", "WMB/USD", "WMT/USD",
                    "WUY/USD", "WYN/USD", "WYY/USD", "XOM/USD", "XRA/USD", "XRX/USD",
                    "XYY/USD", "YUM/USD", "YYY/USD",
                 'SUY/USD','IVV/USD','XLF/USD','KBR/USD','MOH/USD','AZP/USD','WAT/USD',
                 'FMX/USD','FTE/USD','ABB/USD','SYK/USD','LEC/USD','PAR/USD','BID/USD',
                 'KIM/USD','RCL/USD','VXF/USD','INC/USD','DLR/USD','MLM/USD','RYA/USD',
                 'SHA/USD','VTV/USD','VRT/USD','KRE/USD','MAC/USD','ANS/USD','ITT/USD',
                 'HCA/USD','ALS/USD','SAB/USD','OGE/USD','EPD/USD','PPL/USD','BCE/USD',
                 'NWL/USD','LIT/USD','NRG/USD','PII/USD','PTC/USD','SQY/USD','IWB/USD',
                 'IRM/USD','JBL/USD','BMR/USD','RNG/USD','AMX/USD','INF/USD','UDR/USD',
                 'ATR/USD','TSC/USD','ORI/USD','THO/USD','JEF/USD','PSE/USD','MEL/USD',
                 'GAT/USD','FBI/USD','FRT/USD','OSK/USD','CCJ/USD','CDN/USD','JKS/USD',
                 'LKQ/USD','EQI/USD','HYG/USD','BXY/USD','TDO/USD','TXR/USD','TPH/USD',
                 'CAH/USD','DGR/USD','EFA/USD','AZN/USD','BAM/USD','VNQ/USD','KMI/USD',
                 'LHY/USD','HES/USD','AVB/USD','SMG/USD','GEN/USD','HEI/USD','RNR/USD',
                 'KMT/USD','IMO/USD','VWO/USD','COM/USD','GDX/USD','BOX/USD','CYB/USD',
                 'MCK/USD','VOY/USD','GFF/USD','BAP/USD','APD/USD','CHI/USD','VYX/USD',
                 'EMN/USD','ATI/USD','SNP/USD','GNT/USD','VGT/USD','OKE/USD','ZBR/USD',
                 'BSA/USD','SOH/USD','LQD/USD','AEM/USD','RSY/USD','TXT/USD','FDN/USD',
                 'MPW/USD','CRL/USD','NXP/USD','PHM/USD','MAN/USD','MTD/USD','BPY/USD',
                 'KOF/USD','SPL/USD','OLE/USD','VAL/USD','WMY/USD','ZYY/USD','PFG/USD',
                 'PWR/USD','XLU/USD','SON/USD','TEC/USD','HUN/USD','TRM/USD','ETS/USD',
                 'CCI/USD','VHT/USD','AEO/USD','BRY/USD','BKN/USD','VIA/USD','VYM/USD',
                 'MSC/USD','TMO/USD','CSI/USD','ETN/USD','SYF/USD','ALB/USD','LPL/USD',
                 'MRV/USD','HST/USD','CTL/USD','BGY/USD','RSP/USD','SJM/USD','TIP/USD',
                 'VIR/USD','XBI/USD','HBI/USD','CMS/USD','BSV/USD','ESY/USD','ICL/USD',
                 'FNB/USD','IWS/USD','KYY/USD','ALR/USD','AWK/USD','VMI/USD','CHK/USD',
                 'HYY/USD','EDY/USD','ENS/USD','TRI/USD','LAZ/USD','FVD/USD','BBD/USD',
                 'VTR/USD','JAZ/USD','BUS/USD','NOW/USD','COL/USD','FIS/USD','LEG/USD',
                 'GNR/USD','DGY/USD','KMX/USD','ENR/USD','WMS/USD','DDS/USD','WPC/USD',
                 'DKS/USD','MDL/USD','SUM/USD','PSA/USD','XLK/USD','FLS/USD','CDE/USD',
                 'QRV/USD','MEO/USD','KNX/USD','XLY/USD','MTG/USD','ARW/USD','BXM/USD',
                 'GOL/USD','PAA/USD','MKL/USD','XLE/USD','GPR/USD','SXC/USD','CMI/USD',
                 'ATO/USD','ITY/USD','RFY/USD','AMC/USD','HIG/USD','URI/USD','BWA/USD',
                 'ENP/USD','SLG/USD','HMC/USD','XLI/USD','NEU/USD','EVT/USD','HII/USD',
                 'SRC/USD','AON/USD','REG/USD','TTW/USD','GLD/USD','TTE/USD','PNR/USD',
                 'CDW/USD','YEL/USD','MDU/USD','PHY/USD','AMB/USD','BLD/USD','MGA/USD',
                 'AME/USD','HPY/USD','PNW/USD','CNO/USD','ELV/USD','ASX/USD','TEO/USD',
                 'PVH/USD','COR/USD','CLF/USD','AGG/USD','XLP/USD','IBN/USD','LNC/USD',
                 'OMC/USD','ZBH/USD','VPL/USD','VMC/USD','DLB/USD','ITW/USD','VIG/USD',
                 'IWM/USD','ITU/USD','TEA/USD','TBT/USD','MSM/USD','LUV/USD','JWN/USD',
                 'TPR/USD','ILM/USD','AMH/USD','CBO/USD','ADI/USD','AMT/USD','LPG/USD',
                 'SQM/USD','TER/USD','VUG/USD','TDY/USD','ECY/USD','CAG/USD','CCK/USD',
                 'CPT/USD','HIM/USD','LBT/USD','MTB/USD','CTR/USD','ORL/USD','CHT/USD',
                 'ITR/USD','GDD/USD','MDY/USD','MPL/USD','VBY/USD','LUL/USD','BKU/USD',
                 'VEA/USD','DIA/USD','LXP/USD','IWO/USD','RYN/USD','ETR/USD','PLD/USD',
                 'AGO/USD','NWS/USD','OEF/USD','CVI/USD','SYN/USD','WIX/USD','CAK/USD',
                 'SPD/USD','TEV/USD','ESS/USD','CFG/USD','HXL/USD','WPM/USD','ARY/USD',
                 'CRT/USD','NDS/USD','DVA/USD','WDA/USD','NYT/USD','DEI/USD','CLX/USD',
                 'AGC/USD','ONY/USD','GPC/USD','AEE/USD','EIX/USD','GWW/USD','COT/USD',
                 'CLH/USD','RRC/USD','HOU/USD','NVS/USD','THC/USD','ADP/USD','CEY/USD',
                 'AGN/USD','AVT/USD','AYI/USD','GPN/USD','MPC/USD','EQR/USD','OLL/USD',
                 'HDV/USD','MKT/USD','LHX/USD','TSN/USD','TMY/USD','LNG/USD','CPA/USD',
                 'VTI/USD','IFF/USD','VFC/USD','SAN/USD','XLB/USD','TGI/USD','STL/USD',
                 'FIV/USD','MTC/USD','TRO/USD','BAH/USD','XLV/USD','HAS/USD','DAN/USD',
                 'FLY/USD','DOV/USD','IVZ/USD','OYY/USD','UNM/USD','BCY/USD','ETD/USD',
                 'DHR/USD','AFG/USD','GFI/USD','TSM/USD','WCC/USD','STZ/USD','FSL/USD',
                 'NGD/USD','HPE/USD','PNN/USD','TFC/USD','SFM/USD','CRU/USD','BLM/USD',
                 'PRG/USD','IQV/USD','RAC/USD','UIY/USD','LYB/USD','WEC/USD','HBA/USD',
                 'FFI/USD','SPR/USD','OCY/USD','RGL/USD','HUB/USD','POO/USD','CCU/USD',
                 'ROP/USD','NAV/USD','VTY/USD','HRB/USD','TAP/USD','AJG/USD','DLT/USD',
                 'AYY/USD','IVE/USD','CXY/USD','BDX/USD','FLE/USD','PYP/USD','WDC/USD',
                 'TFX/USD','RUN/USD','IWD/USD','VEU/USD','HMY/USD','ZIO/USD','EXR/USD',
                 'TCO/USD','MLC/USD','ALK/USD','ACN/USD','FNF/USD','BYO/USD','TRN/USD',
                 'VOO/USD','ACM/USD','PSX/USD','TGN/USD','ELY/USD','IPG/USD','IDX/USD',
                 'JBH/USD','ATH/USD','PPG/USD','URB/USD','MAS/USD','KRC/USD','VDE/USD',
                 'RVT/USD','IEM/USD','WLK/USD','ALG/USD','KRY/USD','SLM/USD','FHN/USD',
                 'CVL/USD','TEL/USD','WRB/USD','IEP/USD','ECL/USD','NBI/USD','KKR/USD',
                 'UMC/USD','GLO/USD','ZTS/USD','WKC/USD','MOS/USD','AZT/USD','SWN/USD',
                 'BIO/USD','EHT/USD','IWN/USD','ERI/USD','MFC/USD','BFH/USD','IGT/USD',
                 'WBY/USD','MSG/USD','SPY/USD','CRY/USD','LPX/USD','TNE/USD','FLO/USD',
                 'SPW/USD','BRF/USD','GSL/USD','AGY/USD','CLY/USD','HOL/USD','JDY/USD',
                 'CIB/USD','AER/USD','EEM/USD','GTY/USD','JYY/USD','WTW/USD','XEL/USD',
                 'LYV/USD']

# UK_stocks = ['AAL/USD','ABF/USD',
# 'ADM/USD',
# 'AHT/USD','AML/USD',
# 'ANT/USD',
# 'ASC/USD','ASH/USD','AUT/USD','AZN/USD','BAB/USD','BAR/USD','BAT/USD',
# 'BBY/USD','BHP/USD','BKG/USD','BLN/USD','BNZ/USD',
# 'BRB/USD',
# 'CCH/USD','CCL/USD','CNA/USD','CPG/USD','CPI/USD',
# 'CRD/USD',
# 'CRH/USD','CSP/USD','DCC/USD','DGE/USD','DLG/USD','DNL/USD','DOM/USD','DRX/USD',
# 'ELM/USD',
# 'EMG/USD','EXP/USD','EZJ/USD','FER/USD','FGP/USD',
# 'FRE/USD','GLE/USD','GNC/USD','GRG/USD','GSK/USD','HAS/USD','HIK/USD',
#              'HLM/USD','HMS/USD','HOC/USD','HSB/USD','HSX/USD','HWD/USD','IAG/USD','IGG/USD',
# 'IHG/USD',
# 'III/USD','IMB/USD','INC/USD','INF/USD','INV/USD','IPO/USD','ITR/USD','ITV/USD',
#              'JMA/USD','JUP/USD','JUS/USD','KGF/USD','LAN/USD','LGE/USD','LLO/USD','MAB/USD','MAR/USD','MKS/USD','MND/USD','MON/USD','MRO/USD',
# 'MTO/USD','MTR/USD','NXT/USD','OCD/USD','PAG/USD',
# 'PAE/USD',
# 'PAY/USD','PET/USD','PHN/USD',
# 'PLU/USD','PNN/USD','PRU/USD','PSN/USD','PSO/USD',
# 'REL/USD','RIO/USD','RMV/USD','ROR/USD','RTO/USD','SBR/USD','SDR/USD','SGE/USD','SGR/USD','SHI/USD','SMI/USD','SMT/USD',
# 'SPT/USD',
# 'SPX/USD','SSE/USD','STJ/USD','SVT/USD','TLW/USD',
# 'TPK/USD',
# 'TSC/USD',
# 'ULV/USD','VOD/USD','VSV/USD','WEI/USD','WIZ/USD',
# 'WPP/USD',
# 'WTB/USD','3IN/USD','AGR/USD','AGT/USD','ALF/USD','ALW/USD','ASL/USD','ATG/USD','ATY/USD','AVO/USD',
#     'BAG/USD','BBO/USD','BEZ/USD','BGE/USD','BHM/USD','BIR/USD','BME/USD','BNK/USD','BOY/USD',
#     'BRE/USD','BTR/USD','BWY/USD','BYG/USD','CAR/USD','CBG/USD','CCC/USD','CCR/USD','CKN/USD',
#     'CLD/USD','COA/USD','CRS/USD','CTE/USD','CTY/USD','CWK/USD','DEC/USD','DFS/USD','DLN/USD',
#     'DOC/USD','DPL/USD','DWL/USD','EDI/USD','EDV/USD','ENG/USD','ENO/USD','ESN/USD','FCH/USD',
#     'FCI/USD','FCS/USD','FEV/USD','FGT/USD','FNC/USD','FXP/USD','GAW/USD','GEN/USD','GFR/USD',
#     'GFT/USD','GKP/USD','GNS/USD','GPE/USD','GRI/USD','HBR/USD','HIC/USD','HIL/USD','HLN/USD',
#     'HLT/USD','HTG/USD','HTW/USD','HVP/USD','IBS/USD','ICG/USD','IMI/USD','IND/USD','INP/USD',
#     'IWG/USD','JAM/USD','JDW/USD','JMG/USD','KIE/USD','KLR/USD','LMP/USD','LRE/USD','LWD/USD',
#     'MCG/USD','MGA/USD','MNK/USD','MOO/USD','MRC/USD','MRH/USD','MSL/USD','MYI/USD','NCC/USD',
#     'NRR/USD','OMU/USD','ONT/USD','OSB/USD','OXB/USD','PCT/USD','PNL/USD','PTE/USD','PZC/USD',
#     'QLT/USD','RAT/USD','RCP/USD','RNK/USD','RSW/USD','SAF/USD','SCT/USD','SEQ/USD','SHC/USD',
#     'SHE/USD','SMW/USD','SNR/USD','SPI/USD','SRP/USD','SSP/USD','SUK/USD','SUP/USD','SVS/USD',
#     'SWR/USD','SXS/USD','SYN/USD','TAT/USD','TBC/USD','TEM/USD','TEP/USD','TMP/USD','TRI/USD',
#     'TRN/USD','TRY/USD','UKW/USD','UTG/USD','WIS/USD','WKP/USD','WOS/USD','WWH/USD','ZIG/USD']

# Netherland_stocks = ['AAL/USD','ABN/USD','ADX/USD','ADY/USD','AGN/USD','AKZ/USD','APA/USD','ARC/USD','ASX/USD',
    # 'ASM/USD','ASR/USD','BAM/USD','BES/USD','CRB/USD','ECM/USD','GLP/USD','HEI/USD','IMC/USD',
    # 'ING/USD','KPN/USD','LIG/USD','MTY/USD','NNY/USD','PHI/USD','PNL/USD','RAN/USD','SBM/USD',
    # 'TKW/USD','TOM/USD','TWE/USD','UNA/USD','VPK/USD','WHA/USD','WKL/USD']

# France_stocks = ['ACY/USD','ACA/USD','ADP/USD','AFY/USD','AIY/USD','AIR/USD','AKE/USD','ALO/USD','ALT/USD',
#     'AMU/USD','BIM/USD','BNY/USD','BNP/USD','BOL/USD','BVI/USD','CAY/USD','CAP/USD','CDI/USD',
#     'COV/USD','CSY/USD','DEC/USD','DGY/USD','DIM/USD','DSY/USD','EDE/USD','ELY/USD','ELI/USD',
#     'ENY/USD','ENG/USD','ERF/USD','ETL/USD','FGR/USD','FRY/USD','FRV/USD','GET/USD','GLE/USD',
#     'HOY/USD','ICA/USD','IPN/USD','KER/USD','LIY/USD','LRY/USD','MCY/USD','MER/USD','MFY/USD',
#     'MLY/USD','MMB/USD','MMT/USD','NEX/USD','NKY/USD','ORY/USD','ORA/USD','PUB/USD','RCO/USD',
#     'RIY/USD','RMS/USD','RNO/USD','RUI/USD','RXL/USD','SAF/USD','SAN/USD','SCR/USD','SES/USD',
#     'SGO/USD','SUY/USD','SWY/USD','TEP/USD','TFI/USD','UBI/USD','URW/USD','VIE/USD','VIV/USD',
#     'WLN/USD']
# Belgium_stocks = ['ABI/USD','ARG/USD','BEK/USD','BPO/USD','COF/USD','DIE/USD','GBL/USD','KBC/USD',
#     'SOL/USD','TES/USD','UCB/USD','UMI/USD']

# Portugal_stocks = ['CTT/USD','EDP/USD','GAL/USD','JMT/USD','SON/USD']

# CMC_US_Equities = ["AAO/USD", "AAT/USD", "ABE/USD", "ABB/USD", "ABT/USD", "ABG/USD", "ABM/USD", "ABX/USD", "ACH/USD", "ACC/USD", "ACM/USD", "ACI/USD", "ACN/USD", "AC1/USD", "ACP/USD", "ACT/USD", "ADB/USD", "ADM/USD", "ACU/USD", "ACX/USD",
#     "ADC/USD", "AD1/USD", "ADI/USD", "ADN/USD", "ADT/USD", "ADP/USD", "ADS/USD", "ADU/USD", "AEI/USD", "ADX/USD", "AEE/USD", "AEG/USD", "AEO/USD", "AEM/USD", "AES/USD", "AEP/USD", "AFB/USD", "AER/USD", "AFG/USD", "AGI/USD",
#     "AGC/USD", "AGN/USD", "AG1/USD", "AGY/USD", "AGX/USD", "AGM/USD", "AGR/USD", "AG2/USD", "AHH/USD", "AIG/USD", "AIR/USD", "AI1/USD", "AIN/USD", "AKA/USD", "AIZ/USD", "AIT/USD", "AKR/USD", "AJG/USD", "ALB/USD", "ALC/USD",
#     "ALE/USD", "ALG/USD", "AL1/USD", "AL2/USD", "ALK/USD", "AL3/USD", "ALL/USD", "AL4/USD", "AL5/USD", "AL6/USD", "ALR/USD", "AL7/USD", "ALO/USD", "ALV/USD", "ALS/USD", "AMB/USD", "ALX/USD", "AMA/USD", "AM1/USD", "AME/USD",
#     "AM2/USD", "AMC/USD", "AMK/USD", "AMG/USD", "AMN/USD", "AMP/USD", "AMH/USD", "AM3/USD", "AM4/USD", "AMR/USD", "AMS/USD", "AMW/USD", "AMT/USD", "AMX/USD", "AND/USD", "AM5/USD", "ANI/USD", "ANF/USD", "ANG/USD", "ANS/USD",
#     "AON/USD", "ANX/USD", "AOS/USD", "AOD/USD", "APD/USD", "APL/USD", "APO/USD", "APP/USD", "AP1/USD", "APT/USD", "ARC/USD", "AP2/USD", "AR1/USD", "AR2/USD", "ARD/USD", "ARE/USD", "ARO/USD", "ART/USD", "ASA/USD", "ARW/USD",
#     "ARX/USD", "AR3/USD", "ASB/USD", "ASC/USD", "ASG/USD", "ASH/USD", "ASR/USD", "AS1/USD", "ASI/USD", "ASM/USD", "ASX/USD", "ASU/USD", "ASY/USD", "ATI/USD", "ATL/USD", "ATE/USD", "ATG/USD", "ATH/USD", "ATR/USD", "ATO/USD",
#     "AUD/USD", "AUX/USD", "AUB/USD", "AT1/USD", "AT2/USD", "AVA/USD", "AVD/USD", "AV1/USD", "AVN/USD", "AVK/USD", "AV2/USD", "AV3/USD", "AVT/USD", "AVY/USD", "AWF/USD", "AWP/USD", "AWR/USD", "AWK/USD", "AWI/USD", "AW1/USD",
#     "AXL/USD", "AXG/USD", "AXO/USD", "AXP/USD", "AXT/USD", "AXX/USD", "AYI/USD", "AZT/USD", "AZN/USD", "AX1/USD", "BAM/USD", "AZZ/USD", "BAH/USD", "BAN/USD", "BAP/USD", "BA1/USD", "BAT/USD", "BA2/USD", "BA3/USD", "BAX/USD",
#     "BBD/USD", "BB1/USD", "BBN/USD", "BBY/USD", "BCB/USD", "BCC/USD", "BBV/USD", "BCO/USD", "BCE/USD", "BCS/USD", "BCP/USD", "BDX/USD", "BDJ/USD", "BCX/USD", "BDC/USD", "BEL/USD", "BDN/USD", "BEN/USD", "BFA/USD", "BEP/USD",
#     "BF./USD", "BFK/USD", "BFI/USD", "BF1/USD", "BGB/USD", "BFS/USD", "BGS/USD", "BGH/USD", "BGR/USD", "BG1/USD", "BGT/USD", "BGX/USD", "BGY/USD", "BHB/USD", "BG2/USD", "BHF/USD", "BHK/USD", "BHL/USD", "BHE/USD", "BHP/USD",
#     "BIO/USD", "BID/USD", "BIP/USD", "BIT/USD", "BJR/USD", "BKN/USD", "BKE/USD", "BKD/USD", "BK1/USD", "BKR/USD", "BKX/USD", "BKU/USD", "BLB/USD", "BLD/USD", "BLE/USD", "BL1/USD", "BLM/USD", "BL2/USD", "BLK/USD", "BLW/USD",
#     "BL3/USD", "BLX/USD", "BMO/USD", "BL4/USD", "BME/USD", "BMI/USD", "BMY/USD", "BMR/USD", "BNS/USD", "BOE/USD", "BOK/USD", "BOO/USD", "BOH/USD", "BO1/USD", "BPO/USD", "BOX/USD", "BPM/USD", "BPR/USD", "BRC/USD", "BPX/USD",
#     "BPT/USD", "BRF/USD", "BRK/USD", "BR1/USD", "BR2/USD", "BRT/USD", "BRX/USD", "BR3/USD", "BR4/USD", "BSB/USD", "BSA/USD", "BSL/USD", "BST/USD", "BSE/USD", "BSX/USD", "BSR/USD", "BTA/USD", "BTG/USD", "BTO/USD", "BTZ/USD",
#     "BUD/USD", "BUI/USD", "BTT/USD", "BWA/USD", "BVN/USD", "BUS/USD", "BUR/USD", "BWG/USD", "BWF/USD", "BWX/USD", "BZH/USD", "BXM/USD", "BYX/USD", "BYM/USD", "BXX/USD", "BZU/USD", "CAC/USD", "CA1/USD", "CA2/USD", "CAD/USD",
#     "CAF/USD", "CAG/USD", "CAH/USD", "CAK/USD", "CAL/USD", "CA3/USD", "CAM/USD", "CAS/USD", "CAT/USD", "CA4/USD", "CAR/USD", "CA5/USD", "CA6/USD", "CBA/USD", "CBS/USD", "CBR/USD", "CBT/USD", "CB1/USD", "CBZ/USD", "CBU/USD",
#     "CCB/USD", "CBX/USD", "CCE/USD", "CCK/USD", "CCJ/USD", "CCI/USD", "CCO/USD", "CCX/USD", "CCU/USD", "CCN/USD", "CCS/USD", "CDE/USD", "CDW/USD", "CDN/USD", "CDX/USD", "CEE/USD", "CEF/USD", "CDZ/USD", "CEN/USD", "CE1/USD",
#     "CEX/USD", "CER/USD", "CFF/USD", "CEV/USD", "CFG/USD", "CFX/USD", "CFR/USD", "CF1/USD", "CGB/USD", "CGX/USD", "CHD/USD", "CHC/USD", "CH1/USD", "CHE/USD", "CHH/USD", "CHG/USD", "CHN/USD", "CHI/USD", "CHK/USD", "CHM/USD",
#     "CHR/USD", "CHW/USD", "CHT/USD", "CH2/USD", "CIA/USD", "CHY/USD", "CIG/USD", "CII/USD", "CI1/USD", "CIE/USD", "CIO/USD", "CIN/USD", "CIV/USD", "CLB/USD", "CLD/USD", "CLF/USD", "CIX/USD", "CLA/USD", "CLH/USD", "CLM/USD",
#     "CL1/USD", "CLP/USD", "CLX/USD", "CLS/USD", "CLW/USD", "CL2/USD", "CMC/USD", "CM1/USD", "CME/USD", "CMA/USD", "CM2/USD", "CMP/USD", "CM3/USD", "CMR/USD", "CMI/USD", "CM4/USD", "CMX/USD", "CMS/USD", "CNC/USD", "CNA/USD",
#     "CNN/USD", "CNP/USD", "CNI/USD", "CNM/USD", "CNO/USD", "CN1/USD", "CNS/USD", "CNR/USD", "CNX/USD", "CN2/USD", "COH/USD", "COF/USD", "COL/USD", "CO1/USD", "CO2/USD", "CO3/USD", "CO4/USD", "COS/USD", "COP/USD", "COR/USD",
#     "COT/USD", "CPA/USD", "CPF/USD", "CRD/USD", "CPT/USD", "CPK/USD", "CPB/USD", "CRA/USD", "CRH/USD", "CR1/USD", "CRE/USD", "CRI/USD", "CRM/USD", "CRL/USD", "CR2/USD", "CRO/USD", "CRS/USD", "CRU/USD", "CR3/USD", "CRT/USD",
#     "CRX/USD", "CRW/USD", "CSG/USD", "CSC/USD", "CST/USD", "CSI/USD", "CSL/USD", "CSQ/USD", "CSV/USD", "CS1/USD", "CTR/USD", "CTB/USD", "CSW/USD", "CTS/USD", "CT1/USD", "CUK/USD", "CT2/USD", "CUB/USD", "CVB/USD", "CVC/USD",
#     "CUL/USD", "CVG/USD", "CVL/USD", "CVX/USD", "CVS/USD", "CVI/USD", "CWB/USD", "CWC/USD", "CWX/USD", "CWH/USD", "CXE/USD", "CXX/USD", "CYD/USD", "CWT/USD", "CX1/USD", "CYB/USD", "CZN/USD", "DAR/USD", "CZW/USD", "DAN/USD",
#     "DBL/USD", "DBX/USD", "DCO/USD", "DC1/USD", "DDS/USD", "DEI/USD", "DCI/USD", "DEO/USD", "DEX/USD", "DEN/USD", "DFP/USD", "DFI/USD", "DGI/USD", "DGX/USD", "DG1/USD", "DHF/USD", "DG2/USD", "DHI/USD", "DHR/USD", "DH1/USD",
#     "DHX/USD", "DHY/USD", "DHT/USD", "DIS/USD", "DIO/USD", "DLR/USD", "DLH/USD", "DLB/USD", "DKS/USD", "DLT/USD", "DL1/USD", "DMB/USD", "DMO/USD", "DLX/USD", "DNN/USD", "DNL/USD", "DNO/USD", "DNP/USD", "DOC/USD", "DOR/USD",
#     "DPG/USD", "DPZ/USD", "DRI/USD", "DOX/USD", "DSG/USD", "DSL/USD", "DSM/USD", "DTE/USD", "DSX/USD", "DUK/USD", "DXC/USD", "DVA/USD", "DVN/USD", "DYX/USD", "DXX/USD", "EAD/USD", "EBM/USD", "EAT/USD", "EAX/USD", "EBA/USD",
#     "EBF/USD", "ECL/USD", "ECP/USD", "ECV/USD", "ECX/USD", "EEF/USD", "EDU/USD", "EEX/USD", "EDX/USD", "EDD/USD", "EFR/USD", "EFT/USD", "EFX/USD", "EGH/USD", "EGA/USD", "EGP/USD", "EGB/USD", "EHC/USD", "EGY/USD", "EIM/USD",
#     "EIG/USD", "ELF/USD", "EHI/USD", "EIX/USD", "EHT/USD", "EMD/USD", "ELX/USD", "EME/USD", "EMN/USD", "EMF/USD", "EMR/USD", "ENB/USD", "ENI/USD", "ENR/USD", "ENP/USD", "ENS/USD", "ENT/USD", "ENV/USD", "EOD/USD", "EN1/USD",
#     "EOG/USD", "EPA/USD", "EOI/USD", "EOT/USD", "EP1/USD", "EOS/USD", "EPC/USD", "EPD/USD", "EPM/USD", "EQI/USD", "EQT/USD", "EQB/USD", "EQN/USD", "EPR/USD", "EQR/USD", "ERI/USD", "ERC/USD", "ESC/USD", "ER1/USD", "ESE/USD",
#     "ER2/USD", "ERJ/USD", "ESI/USD", "ESR/USD", "ESN/USD", "ESL/USD", "ESS/USD", "ES1/USD", "ESX/USD", "ETB/USD", "ETD/USD", "ETG/USD", "ETS/USD", "ETR/USD", "ETN/USD", "ETV/USD", "ETJ/USD", "ETW/USD", "ETX/USD", "ETY/USD",
#     "ET1/USD", "EVG/USD", "EVH/USD", "EVN/USD", "EVF/USD", "EVI/USD", "EVM/USD", "EVR/USD", "EVT/USD", "EVV/USD", "EV1/USD", "EXG/USD", "EWB/USD", "EXA/USD", "EXK/USD", "EXE/USD", "EXC/USD", "EXP/USD", "EX1/USD", "EXR/USD",
#     "EX2/USD", "EZP/USD", "EYE/USD", "FAF/USD", "EXX/USD", "FBI/USD", "FBK/USD", "FBP/USD", "FCC/USD", "FCF/USD", "FCB/USD", "FC1/USD", "FCN/USD", "FCP/USD", "FCX/USD", "FCT/USD", "FDS/USD", "FDP/USD", "FEI/USD", "FC2/USD",
#     "FDX/USD", "FEL/USD", "FFA/USD", "FEX/USD", "FFB/USD", "FFC/USD", "FFI/USD", "FF1/USD", "FF2/USD", "FFX/USD", "FHI/USD", "FHN/USD", "FHB/USD", "FGB/USD", "FIS/USD", "FIB/USD", "FI1/USD", "FIC/USD", "FIT/USD", "FIV/USD",
#     "FI2/USD", "FLC/USD", "FLE/USD", "FIX/USD", "FLI/USD", "FLO/USD", "FLW/USD", "FLG/USD", "FLX/USD", "FLS/USD", "FMA/USD", "FMB/USD", "FMN/USD", "FMX/USD", "FMS/USD", "FMC/USD", "FND/USD", "FNL/USD", "FNB/USD", "FNF/USD",
#     "FNV/USD", "FNW/USD", "FNX/USD", "FOF/USD", "FOR/USD", "FOL/USD", "FON/USD", "FO1/USD", "FO2/USD", "FOX/USD", "FO3/USD", "FPF/USD", "FO4/USD", "FPH/USD", "FPI/USD", "FRS/USD", "FRB/USD", "FRM/USD", "FRA/USD", "FRP/USD",
#     "FSM/USD", "FSB/USD", "FRX/USD", "FRT/USD", "FSL/USD", "FSP/USD", "FSS/USD", "FST/USD", "FTA/USD", "FTF/USD", "FTI/USD", "FUL/USD", "FU1/USD", "FTV/USD", "FUN/USD", "FWO/USD", "GAB/USD", "FU2/USD", "FWR/USD", "GAI/USD",
#     "GA1/USD", "GA2/USD", "GAM/USD", "GAS/USD", "GAT/USD", "GBC/USD", "GDD/USD", "GBA/USD", "GBD/USD", "GDO/USD", "GD1/USD", "GDX/USD", "GEL/USD", "GEF/USD", "GFF/USD", "GEN/USD", "GFX/USD", "GE1/USD", "GFI/USD", "GGB/USD",
#     "GGN/USD", "GGT/USD", "GHM/USD", "GHY/USD", "GIC/USD", "GHC/USD", "GIB/USD", "GIL/USD", "GIF/USD", "GI1/USD", "GI2/USD", "GKO/USD", "GLD/USD", "GLO/USD", "GL1/USD", "GLP/USD", "GIS/USD", "GL2/USD", "GLV/USD", "GLQ/USD",
#     "GLW/USD", "GMR/USD", "GMS/USD", "GME/USD", "GNE/USD", "GMX/USD", "GNS/USD", "GNR/USD", "GNT/USD", "GN1/USD", "GN2/USD", "GOF/USD", "GNW/USD", "GOL/USD", "GOO/USD", "GO1/USD", "GPC/USD", "GPI/USD", "GPR/USD", "GRF/USD",
#     "GPK/USD", "GP1/USD", "GRB/USD", "GPN/USD", "GRM/USD", "GRX/USD", "GSB/USD", "GSX/USD", "GTE/USD", "GSK/USD", "GTN/USD", "GTL/USD", "GTY/USD", "GTX/USD", "GVA/USD", "GUT/USD", "GWR/USD", "GWW/USD", "GW1/USD", "GXX/USD",
#     "HAI/USD", "HAF/USD", "HAE/USD", "HAL/USD", "HAS/USD", "HA1/USD", "HBA/USD", "HBI/USD", "HA2/USD", "HBC/USD", "HBB/USD", "HBM/USD", "HCC/USD", "HCI/USD", "HCK/USD", "HCS/USD", "HCA/USD", "HCM/USD", "HDB/USD", "HDX/USD",
#     "HEL/USD", "HEI/USD", "HEX/USD", "HGV/USD", "HII/USD", "HEQ/USD", "HIO/USD", "HIM/USD", "HIW/USD", "HIX/USD", "HLI/USD", "HLX/USD", "HI1/USD", "HLN/USD", "HMS/USD", "HMC/USD", "HMY/USD", "HL1/USD", "HNR/USD", "HOG/USD",
#     "HNI/USD", "HOL/USD", "HON/USD", "HOF/USD", "HPE/USD", "HOP/USD", "HO1/USD", "HPF/USD", "HPI/USD", "HPQ/USD", "HPS/USD", "HPP/USD", "HPX/USD", "HQH/USD", "HRB/USD", "HQY/USD", "HQL/USD", "HRL/USD", "HRT/USD", "HSI/USD",
#     "HRZ/USD", "HRX/USD", "HST/USD", "HSB/USD", "HS1/USD", "HS2/USD", "HSY/USD", "HTG/USD", "HTD/USD", "HTH/USD", "HUB/USD", "HTL/USD", "HU1/USD", "HUR/USD", "HU2/USD", "HUN/USD", "HVT/USD", "HWB/USD", "HUM/USD", "HXL/USD",
#     "HWC/USD", "HYT/USD", "HYI/USD", "HYX/USD", "HZO/USD", "HXX/USD", "IAG/USD", "IBO/USD", "IBC/USD", "IBM/USD", "IAR/USD", "IBN/USD", "IBP/USD", "ICH/USD", "ICU/USD", "ICF/USD", "ICL/USD", "IDC/USD", "IDA/USD", "IC1/USD",
#     "IEX/USD", "IDE/USD", "IES/USD", "IEP/USD", "IGA/USD", "IFN/USD", "IFF/USD", "IGD/USD", "IGI/USD", "IHD/USD", "IGR/USD", "IIP/USD", "IIF/USD", "III/USD", "II1/USD", "IIM/USD", "IHG/USD", "IMO/USD", "ILM/USD", "ILP/USD",
#     "IM1/USD", "INC/USD", "INB/USD", "IMK/USD", "INN/USD", "INF/USD", "IN1/USD", "IND/USD", "ING/USD", "INS/USD", "IN2/USD", "IN3/USD", "INT/USD", "INV/USD", "IN4/USD", "ION/USD", "IOS/USD", "IPA/USD", "IN5/USD", "IN6/USD",
#     "IPG/USD", "IQV/USD", "IP1/USD", "IRM/USD", "IQI/USD", "IPX/USD", "IRD/USD", "IR1/USD", "ISD/USD", "IRS/USD", "IST/USD", "IRW/USD", "ISS/USD", "IRT/USD", "ITG/USD", "ITR/USD", "IT1/USD", "ITU/USD", "IVZ/USD", "ITW/USD",
#     "ITX/USD", "JAC/USD", "JBL/USD", "JAZ/USD", "JBG/USD", "JBH/USD", "JCE/USD", "JB1/USD", "JBS/USD", "JDX/USD", "JFR/USD", "JGH/USD", "JEL/USD", "JCI/USD", "JJS/USD", "JHG/USD", "JKH/USD", "JLL/USD", "JLS/USD", "JOE/USD",
#     "JOU/USD", "JPI/USD", "JPC/USD", "JOF/USD", "JRS/USD", "JRI/USD", "JRV/USD", "JYN/USD", "KAL/USD", "JQC/USD", "KBH/USD", "KBR/USD", "KEX/USD", "KEL/USD", "KEP/USD", "KEY/USD", "KE1/USD", "KBX/USD", "KE2/USD", "KFR/USD",
#     "KFY/USD", "KFX/USD", "KIM/USD", "KHC/USD", "KID/USD", "KGC/USD", "KKR/USD", "KIO/USD", "KLI/USD", "KMB/USD", "KMI/USD", "KMD/USD", "KLA/USD", "KMX/USD", "KMP/USD", "KNO/USD", "KMT/USD", "KNX/USD", "KOF/USD", "KN1/USD",
#     "KNS/USD", "KOX/USD", "KOP/USD", "KO1/USD", "KRN/USD", "KRC/USD", "KR1/USD", "KRO/USD", "KRG/USD", "KSS/USD", "KRX/USD", "KTF/USD", "KTC/USD", "KTX/USD", "KTO/USD", "KVH/USD", "KWR/USD", "LAC/USD", "KWX/USD", "KXX/USD",
#     "LAD/USD", "LAK/USD", "LAN/USD", "LAU/USD", "LAM/USD", "LBT/USD", "LAZ/USD", "LCU/USD", "LCN/USD", "LBR/USD", "LB1/USD", "LB2/USD", "LB3/USD", "LDP/USD", "LEC/USD", "LEG/USD", "LEN/USD", "LDO/USD", "LEA/USD", "LE1/USD",
#     "LEO/USD", "LGN/USD", "LFU/USD", "LHX/USD", "LII/USD", "LFV/USD", "LGI/USD", "LH1/USD", "LIL/USD", "LIN/USD", "LI1/USD", "LI2/USD", "LKQ/USD", "LLY/USD", "LIT/USD", "LIV/USD", "LMB/USD", "LMA/USD", "LMT/USD", "LMN/USD",
#     "LNC/USD", "LNG/USD", "LNT/USD", "LNN/USD", "LOB/USD", "LOC/USD", "LOG/USD", "LOW/USD", "LOO/USD", "LPL/USD", "LP1/USD", "LPG/USD", "LPX/USD", "LQD/USD", "LRN/USD", "LSC/USD", "LST/USD", "LUL/USD", "LTC/USD", "LTR/USD",
#     "LVS/USD", "LXF/USD", "LUV/USD", "LWX/USD", "LXX/USD", "LYG/USD", "LZB/USD", "LYT/USD", "LX1/USD", "LYV/USD", "LYB/USD", "MAC/USD", "MAA/USD", "MAI/USD", "MAN/USD", "MAR/USD", "MA1/USD", "MA2/USD", "MAS/USD", "MAV/USD",
#     "MAT/USD", "MAX/USD", "MBW/USD", "MBC/USD", "MA3/USD", "MBU/USD", "MBI/USD", "MCF/USD", "MCD/USD", "MCH/USD", "MCR/USD", "MCK/USD", "MC1/USD", "MCS/USD", "MCY/USD", "MDL/USD", "MDB/USD", "MCX/USD", "MDT/USD", "MED/USD",
#     "MEI/USD", "MDU/USD", "MDX/USD", "MER/USD", "MEL/USD", "MET/USD", "MFC/USD", "ME1/USD", "MFG/USD", "MEO/USD", "MFI/USD", "MFM/USD", "MGA/USD", "MGM/USD", "MGI/USD", "MGF/USD", "MGE/USD", "MGY/USD", "MGR/USD", "MGX/USD",
#     "MGP/USD", "MHI/USD", "MHN/USD", "MHK/USD", "MHD/USD", "MIT/USD", "MHO/USD", "MID/USD", "MIN/USD", "MIY/USD", "MKC/USD", "MKS/USD", "MKL/USD", "MLK/USD", "MLA/USD", "MLC/USD", "MLM/USD", "MMD/USD", "MLP/USD", "MLR/USD",
#     "MMC/USD", "MMS/USD", "MMI/USD", "MMM/USD", "MMT/USD", "MM1/USD", "MND/USD", "MOH/USD", "MOD/USD", "MNR/USD", "MOF/USD", "MOS/USD", "MOR/USD", "MOG/USD", "MPA/USD", "MOX/USD", "MOV/USD", "MPW/USD", "MPC/USD", "MPX/USD",
#     "MP1/USD", "MQT/USD", "MQY/USD", "MRC/USD", "MRK/USD", "MR1/USD", "MSA/USD", "MRU/USD", "MRT/USD", "MSB/USD", "MS1/USD", "MRV/USD", "MSC/USD", "MSG/USD", "MSD/USD", "MSE/USD", "MSX/USD", "MSM/USD", "MTB/USD", "MTD/USD",
#     "MTC/USD", "MTL/USD", "MTG/USD", "MTR/USD", "MT1/USD", "MTS/USD", "MTX/USD", "MTN/USD", "MTZ/USD", "MUA/USD", "MUE/USD", "MVB/USD", "MUS/USD", "MUF/USD", "MUR/USD", "MXF/USD", "MVF/USD", "MVT/USD", "MXL/USD", "MWA/USD",
#     "MYI/USD", "MYR/USD", "MYD/USD", "MYN/USD", "MXX/USD", "MX1/USD", "NAI/USD", "NAC/USD", "NAT/USD", "NAD/USD", "NA1/USD", "NA2/USD", "NAV/USD", "NBB/USD", "NBH/USD", "NBI/USD", "NB1/USD", "NBT/USD", "NBN/USD", "NEA/USD",
#     "NEM/USD", "NDL/USD", "NEO/USD", "NDS/USD", "NE1/USD", "NEU/USD", "NEW/USD", "NEX/USD", "NFL/USD", "NFG/USD", "NGG/USD", "NFJ/USD", "NGD/USD", "NFB/USD", "NGS/USD", "NGV/USD", "NHI/USD", "NGX/USD", "NIC/USD", "NHT/USD",
#     "NHC/USD", "NIX/USD", "NHS/USD", "NJR/USD", "NKS/USD", "NKE/USD", "NKX/USD", "NML/USD", "NMZ/USD", "NMF/USD", "NNI/USD", "NMR/USD", "NNN/USD", "NOA/USD", "NOC/USD", "NOM/USD", "NOV/USD", "NO1/USD", "NRC/USD", "NRG/USD",
#     "NOW/USD", "NRK/USD", "NPK/USD", "NPO/USD", "NRO/USD", "NRI/USD", "NSC/USD", "NSA/USD", "NTA/USD", "NSI/USD", "NTG/USD", "NTC/USD", "NTR/USD", "NTN/USD", "NT1/USD", "NUW/USD", "NUE/USD", "NVE/USD", "NUV/USD", "NVM/USD",
#     "NVO/USD", "NVS/USD", "NVR/USD", "NWB/USD", "NWE/USD", "NWP/USD", "NWS/USD", "NWG/USD", "NWN/USD", "NWF/USD", "NXE/USD", "NXD/USD", "NW1/USD", "NXR/USD", "NXP/USD", "NX1/USD", "NXS/USD", "NYT/USD", "NZF/USD", "NXX/USD",
#     "OCF/USD", "ODF/USD", "OCX/USD", "OFL/USD", "OEC/USD", "OFI/USD", "OFS/USD", "OFG/USD", "OIA/USD", "OGE/USD", "OGS/USD", "OIS/USD", "OLE/USD", "OIX/USD", "OII/USD", "OLL/USD", "OKE/USD", "OLN/USD", "OLP/USD", "ONB/USD",
#     "OMC/USD", "OM1/USD", "OMA/USD", "OMF/USD", "ONT/USD", "ONX/USD", "OOM/USD", "OPY/USD", "ORR/USD", "ORA/USD", "ORI/USD", "ORX/USD", "ORC/USD", "OTE/USD", "OSB/USD", "OSU/USD", "OSK/USD", "OSS/USD", "OSP/USD", "OTT/USD",
#     "OUT/USD", "OZK/USD", "OXX/USD", "OXS/USD", "OXM/USD", "PAG/USD", "PA1/USD", "PAA/USD", "PA2/USD", "PAR/USD", "PAY/USD", "PA3/USD", "PAH/USD", "PA4/USD", "PAT/USD", "PBH/USD", "PBP/USD", "PBR/USD", "PCA/USD", "PCN/USD",
#     "PBT/USD", "PCH/USD", "PCT/USD", "PBX/USD", "PCR/USD", "PCY/USD", "PDE/USD", "PDF/USD", "PDI/USD", "PDL/USD", "PDM/USD", "PEB/USD", "PDT/USD", "PEG/USD", "PE1/USD", "PE2/USD", "PEO/USD", "PEN/USD", "PEP/USD", "PFB/USD",
#     "PES/USD", "PFI/USD", "PFD/USD", "PFG/USD", "PFL/USD", "PGC/USD", "PFS/USD", "PFN/USD", "PGR/USD", "PGP/USD", "PF1/USD", "PG1/USD", "PGX/USD", "PHD/USD", "PHK/USD", "PHI/USD", "PHG/USD", "PHM/USD", "PHT/USD", "PHX/USD",
#     "PII/USD", "PIN/USD", "PJT/USD", "PIM/USD", "PIX/USD", "PIP/USD", "PKG/USD", "PKE/USD", "PKB/USD", "PKX/USD", "PKO/USD", "PLB/USD", "PLD/USD", "PLN/USD", "PLA/USD", "PLO/USD", "PLP/USD", "PML/USD", "PMO/USD", "PLS/USD",
#     "PMF/USD", "PMM/USD", "PLY/USD", "PLX/USD", "PNC/USD", "PNN/USD", "PMX/USD", "PM1/USD", "POD/USD", "PNR/USD", "PNW/USD", "POO/USD", "POR/USD", "PPC/USD", "POS/USD", "PPB/USD", "PPG/USD", "PPI/USD", "PPT/USD", "PRD/USD",
#     "PRA/USD", "PPL/USD", "PR1/USD", "PRG/USD", "PRI/USD", "PR2/USD", "PRK/USD", "PRT/USD", "PSA/USD", "PRU/USD", "PRO/USD", "PR3/USD", "PSF/USD", "PSE/USD", "PSM/USD", "PSO/USD", "PSX/USD", "PTY/USD", "PST/USD", "PTC/USD",
#     "PUK/USD", "PVH/USD", "PWR/USD", "PYP/USD", "PZG/USD", "QLY/USD", "PZZ/USD", "QDE/USD", "QCO/USD", "QNS/USD", "QGE/USD", "QTW/USD", "QRV/USD", "RAC/USD", "QSR/USD", "RAM/USD", "RAI/USD", "RBB/USD", "RAR/USD", "RBC/USD",
#     "RBA/USD", "RB1/USD", "RB2/USD", "RCI/USD", "RCK/USD", "RDI/USD", "RCS/USD", "RDN/USD", "RDC/USD", "RD1/USD", "REG/USD", "RDW/USD", "REI/USD", "RES/USD", "REL/USD", "RFI/USD", "REX/USD", "RE1/USD", "RFX/USD", "RGA/USD",
#     "RGC/USD", "RGL/USD", "RGE/USD", "RGP/USD", "RGR/USD", "RHI/USD", "RGN/USD", "RHX/USD", "RIC/USD", "RHP/USD", "RIO/USD", "RLJ/USD", "RJF/USD", "RLG/USD", "RMD/USD", "RMA/USD", "RMT/USD", "RMR/USD", "RNG/USD", "RMB/USD",
#     "RMX/USD", "RNS/USD", "RNR/USD", "ROC/USD", "ROK/USD", "ROG/USD", "ROL/USD", "ROS/USD", "RO1/USD", "RPD/USD", "ROP/USD", "RQI/USD", "RRX/USD", "RTX/USD", "RRC/USD", "RSX/USD", "RSG/USD", "RUN/USD", "RUS/USD", "RVT/USD",
#     "RVS/USD", "RYA/USD", "RYN/USD", "RYI/USD", "RYT/USD", "RXX/USD", "SAB/USD", "RYX/USD", "SAF/USD", "SAH/USD", "SAC/USD", "SA1/USD", "SAI/USD", "SAN/USD", "SAM/USD", "SA2/USD", "SAP/USD", "SAT/USD", "SBF/USD", "SA3/USD",
#     "SBG/USD", "SBH/USD", "SBA/USD", "SBC/USD", "SBR/USD", "SB1/USD", "SBS/USD", "SB2/USD", "SCC/USD", "SBU/USD", "SCD/USD", "SBX/USD", "SB3/USD", "SCH/USD", "SCI/USD", "SC1/USD", "SCP/USD", "SCM/USD", "SCL/USD", "SCS/USD",
#     "SCV/USD", "SEB/USD", "SC2/USD", "SEI/USD", "SEN/USD", "SED/USD", "SEE/USD", "SEM/USD", "SEV/USD", "SFM/USD", "SFB/USD", "SFL/USD", "SGA/USD", "SFX/USD", "SFS/USD", "SGC/USD", "SGR/USD", "SGU/USD", "SHA/USD", "SHO/USD",
#     "SHG/USD", "SHB/USD", "SH1/USD", "SID/USD", "SIL/USD", "SIG/USD", "SIT/USD", "SI1/USD", "SJT/USD", "SJM/USD", "SKX/USD", "SKT/USD", "SLB/USD", "SLA/USD", "SKY/USD", "SLG/USD", "SLF/USD", "SLM/USD", "SLP/USD", "SLR/USD",
#     "SMF/USD", "SMB/USD", "SM1/USD", "SMG/USD", "SMT/USD", "SMP/USD", "SM2/USD", "SNA/USD", "SNB/USD", "SND/USD", "SNE/USD", "SNF/USD", "SNN/USD", "SNV/USD", "SNP/USD", "SOH/USD", "SNY/USD", "SPG/USD", "SP1/USD", "SON/USD",
#     "SPN/USD", "SOX/USD", "SPO/USD", "SP2/USD", "SPR/USD", "SPX/USD", "SP3/USD", "SRC/USD", "SQM/USD", "SRE/USD", "SRI/USD", "SRD/USD", "SSB/USD", "SSS/USD", "SST/USD", "SS1/USD", "SSD/USD", "STA/USD", "SSY/USD", "ST1/USD",
#     "STE/USD", "STC/USD", "STK/USD", "STL/USD", "STM/USD", "STR/USD", "ST2/USD", "ST3/USD", "ST4/USD", "STN/USD", "ST5/USD", "STW/USD", "STZ/USD", "STX/USD", "SUI/USD", "STT/USD", "ST6/USD", "SUP/USD", "SUX/USD", "SVC/USD",
#     "SWK/USD", "SVM/USD", "SW1/USD", "SWZ/USD", "SWX/USD", "SWB/USD", "SXI/USD", "SYB/USD", "SYK/USD", "SYF/USD", "SYY/USD", "SXT/USD", "SYN/USD", "TAP/USD", "TAC/USD", "TBB/USD", "TCB/USD", "TBP/USD", "TCO/USD", "TCM/USD",
#     "TBI/USD", "TDC/USD", "TDF/USD", "TEA/USD", "TEC/USD", "TEI/USD", "TDO/USD", "TEF/USD", "TEL/USD", "TDX/USD", "TER/USD", "TEV/USD", "TFX/USD", "TFC/USD", "TGN/USD", "TGL/USD", "TGS/USD", "TFS/USD", "TGT/USD", "THG/USD",
#     "THF/USD", "TGX/USD", "THQ/USD", "THR/USD", "TIP/USD", "TH1/USD", "THS/USD", "TIT/USD", "TKC/USD", "TLY/USD", "TKX/USD", "TMP/USD", "TMH/USD", "TKR/USD", "TMO/USD", "TMX/USD", "TMU/USD", "TNE/USD", "TPB/USD", "TNC/USD",
#     "TOW/USD", "TPR/USD", "TOL/USD", "TPH/USD", "TRC/USD", "TRI/USD", "TR1/USD", "TRE/USD", "TRN/USD", "TR2/USD", "TRO/USD", "TRM/USD", "TRP/USD", "TR3/USD", "TR4/USD", "TR5/USD", "TRS/USD", "TRV/USD", "TRU/USD", "TR6/USD",
#     "TSE/USD", "TRX/USD", "TSB/USD", "TS1/USD", "TSI/USD", "TSL/USD", "TSX/USD", "TSQ/USD", "TSN/USD", "TTE/USD", "TTG/USD", "TT1/USD", "TTM/USD", "TWI/USD", "TVT/USD", "TTX/USD", "TTW/USD", "TW1/USD", "TWL/USD", "TVX/USD",
#     "TXN/USD", "TYL/USD", "TXR/USD", "TXX/USD", "TXT/USD", "TYX/USD", "UAA/USD", "TZO/USD", "UAX/USD", "UBS/USD", "UBF/USD", "UB1/USD", "UEI/USD", "UEC/USD", "UDR/USD", "UCT/USD", "UEX/USD", "UFI/USD", "UIX/USD", "UHT/USD",
#     "UHS/USD", "UGI/USD", "ULH/USD", "UFP/USD", "ULT/USD", "ULX/USD", "UNB/USD", "UMB/USD", "UMC/USD", "UNF/USD", "UN1/USD", "UMH/USD", "UNM/USD", "UNH/USD", "UNI/USD", "UPL/USD", "UPS/USD", "UNP/USD", "URI/USD", "URB/USD",
#     "USA/USD", "USB/USD", "US1/USD", "UTF/USD", "USF/USD", "USP/USD", "UTG/USD", "UTH/USD", "UTL/USD", "UTM/USD", "UVE/USD", "UUU/USD", "UVV/USD", "VBF/USD", "UVS/USD", "VAL/USD", "VBT/USD", "VCY/USD", "VCT/USD", "VEC/USD",
#     "VCX/USD", "VEE/USD", "VFC/USD", "VET/USD", "VGZ/USD", "VGI/USD", "VIA/USD", "VGM/USD", "VIC/USD", "VIV/USD", "VIP/USD", "VIR/USD", "VLG/USD", "VLO/USD", "VKQ/USD", "VKI/USD", "VLY/USD", "VMC/USD", "VLR/USD", "VMI/USD",
#     "VND/USD", "VMO/USD", "VNO/USD", "VNR/USD", "VPG/USD", "VOY/USD", "VRA/USD", "VRS/USD", "VOD/USD", "VRN/USD", "VRE/USD", "VRT/USD", "VTR/USD", "VSH/USD", "VR1/USD", "VVR/USD", "VSA/USD", "VST/USD", "VZX/USD", "VVV/USD",
#     "VXX/USD", "WAB/USD", "WAF/USD", "WAS/USD", "WA1/USD", "WBA/USD", "WBX/USD", "WBS/USD", "WDC/USD", "WDA/USD", "WCC/USD", "WDF/USD", "WDX/USD", "WEC/USD", "WEA/USD", "WEN/USD", "WES/USD", "WEX/USD", "WER/USD", "WEY/USD",
#     "WHF/USD", "WGO/USD", "WHG/USD", "WFC/USD", "WIN/USD", "WHR/USD", "WIA/USD", "WIW/USD", "WI1/USD", "WMB/USD", "WLD/USD", "WIX/USD", "WMK/USD", "WKX/USD", "WLK/USD", "WMS/USD", "WNE/USD", "WPC/USD", "WMX/USD", "WNC/USD",
#     "WNS/USD", "WOR/USD", "WPM/USD", "WRB/USD", "WSB/USD", "WPP/USD", "WS1/USD", "WRL/USD", "WSO/USD", "WSC/USD", "WSF/USD", "WSR/USD", "WST/USD", "WTB/USD", "WTF/USD", "WTM/USD", "WTS/USD", "WTR/USD", "WTT/USD", "WYN/USD",
#     "WWD/USD", "WXX/USD", "XEL/USD", "WWW/USD", "WUX/USD", "WYX/USD", "XHR/USD", "XOM/USD", "XNC/USD", "XPE/USD", "XPO/USD", "XYL/USD", "XRA/USD", "YEL/USD", "YOR/USD", "YUM/USD", "YEX/USD", "YU1/USD", "ZBH/USD", "ZDX/USD",
#     "ZTO/USD", "ZEU/USD", "ZIO/USD", "ZBR/USD", "ZGX/USD", "ZTS/USD", "ZUM/USD", "ZWS/USD", "ZXX/USD", "ZYM/USD",]


for file in folders:
    
    new_path = path + file + "/Final Summary/"
    
    for year in years_list:
        
        new_path_1 = new_path + year +"/"
        
        try:
        
            train_df = pd.read_excel(new_path_1 + " Train data summary.xlsx")

            test_df = pd.read_excel(new_path_1 + " Test data summary.xlsx")

            # -------------------- Selection Criteria -------------------- #


            df = train_df.copy()

            # --------- For Best Pairs ----------------- #

            # --------- For Best Pairs ----------------- #

            if year == '2018':
                profit_value = int(round((4/10) * 50000))
                trade_count = int(round((4/10)*100))
            elif year == '2019':
                profit_value = int(round((5/10) * 50000))
                trade_count = int(round((5/10)*100))
            elif year == '2020':
                profit_value = int(round((6/10) * 50000))
                trade_count = int(round((6/10)*100))
            elif year == '2021':
                profit_value = int(round((7/10) * 50000))
                trade_count = int(round((7/10)*100))
            elif year == '2022':
                profit_value = int(round((8/10) * 50000))
                trade_count = int(round((8/10)*100))
            elif year == '2023':
                profit_value = int(round((9/10) * 50000))
                trade_count = int(round((9/10)*100))
            elif year == '2024':
                profit_value = int(round((10/10) * 50000))
                trade_count = int(round((10/10)*100))
            elif year == '2025':
                profit_value = int(round((10/10) * 50000))
                trade_count = int(round((10/10)*100))



            df_1 = df[(df['Total_Trades'] > trade_count) & (df['Net_P_L'] >= profit_value) &
                  (df['% Positive Months by Trade Close Date'] > 0.40)]

            # ---------------- Currency Type ----------------- #

            filters = [df_1['Currency'].isin(Major_pairs_list),
                       df_1['Currency'].isin(Exotic_pairs_list),
                        df_1['Currency'].isin(Commodities_pairs_list),
                        df_1['Currency'].isin(ETF_pairs_list),
                        df_1['Currency'].isin(bond),
                        df_1['Currency'].isin(futures),
                        df_1['Currency'].isin(Index_CFDs),
                        df_1['Currency'].isin(NDFs),
                        df_1['Currency'].isin(Single_Stock_CFDs)]
                        # df_1['Currency'].isin(UK_stocks),
                        # df_1['Currency'].isin(Netherland_stocks),
                        # df_1['Currency'].isin(France_stocks),
                        # df_1['Currency'].isin(Belgium_stocks),
                        # df_1['Currency'].isin(CMC_US_Equities),]
                        # df_1['Currency'].isin(Portugal_stocks)]


            values = ["Non-Exotic",
                      "Exotic",
                      "Commodities",
                     "ETF",
                     "bond","Single Stock CFD","Index CFD","NDF","Single Stock CFD"]


            df_1['Currency Type'] = np.select(filters, values)

            # -------------------------------------------------- #


            df_1["Rank"] = df_1.groupby(by = ["Strategy","Condition","Currency","Timeframe"])["Net_P_L"].rank(ascending=False,method = 'first')

            df_1_rank_1_data = df_1[df_1['Rank'] == 1].reset_index(drop = True)

            # ----------------------------------------- #


            # -------- Rest pairs ---------------- #

            df["Rank"] = df.groupby(by = ["Strategy","Condition","Currency","Timeframe"])["Net_P_L"].rank(ascending=False,method = 'first')

            df_rank_1_data = df[df['Rank'] == 1].reset_index(drop = True)

            # ----------------------------------------- #


            selected_comb_summery = pd.DataFrame(df_1_rank_1_data.groupby(['Currency Type'])['Rank'].count()).reset_index()
            selected_comb_summery['Strategy'] = file

    #         all_stg_selected_comb_list = pd.concat([all_stg_selected_comb_list,selected_comb_summery],ignore_index = True)


            print(file)
            print("Selected combinations", df_1_rank_1_data.groupby(['Currency Type'])['Rank'].count())


            df_1_rank_1_data.to_excel(new_path_1 + "Best combinations BT summary.xlsx",index = False)

            df_rank_1_data.to_excel(new_path_1 + "Rest combinations BT summary.xlsx",index = False)



             # ----------- Extracting Trade level data --------------- #

            all_comb_trade_level_data = pd.DataFrame()

            for row,item in df_1_rank_1_data.iterrows():

                stg = item['Strategy']
                curr = item['Currency']
                curr_1 = curr.replace("/",'_')
                cond = item['Condition']
                tf = item['Timeframe']
                p1 = item['parameter_1']
                p2 = int(item['parameter_2'])
                p3 = item['parameter_3']
                p4 = item['parameter_4']
                p5 = item['parameter_5']
                p6 = item['parameter_6']

                file_name = (curr_1 + "_" + tf + "_" + str(p1) + "_" + str(p2) + "_" +
                             str(p3) + "_" + str(p4) + "_" + str(p5) + "_" + str(p6) + "_" + cond + ".parquet")

        #         print(file_name)
                trade_df = pd.read_parquet(trade_data_path + file + "/Trade Level Data/" + file_name)

                trade_df['Single trade flag'] = np.where(trade_df['date'] < trade_df['Closing Date'].shift(1),1,0)

                trade_df = trade_df[trade_df['Single trade flag'] == 0].reset_index(drop = True)

                all_comb_trade_level_data = pd.concat([all_comb_trade_level_data,trade_df],ignore_index = True)


            # ------------------------------------------------------- #
            
            all_comb_trade_level_data['date'] = pd.to_datetime(all_comb_trade_level_data['date'])
            all_comb_trade_level_data['Closing Date'] = pd.to_datetime(all_comb_trade_level_data['Closing Date'])
            
            all_comb_trade_level_data = all_comb_trade_level_data[all_comb_trade_level_data['date'] >= dt.datetime(2010,1,1,0,0,0)].reset_index(drop = True)


            all_comb_trade_level_data['Single trade flag'] = np.where(all_comb_trade_level_data['date'] < all_comb_trade_level_data['Closing Date'].shift(1),1,0)

            all_comb_trade_level_data = all_comb_trade_level_data[all_comb_trade_level_data['Single trade flag'] == 0].reset_index(drop = True)


            # -------------- Exporting Trades level data -------------- #
            all_comb_trade_level_data.to_csv(new_path_1 + "selected pairs Trade level data.csv",index = False)

        except Exception as e:
            
            print(e)
            
#             break
            
            print("Error")
        
        
        #     print(file)
