==> S3 - Python script to get BT and FT summary.ipynb <==
import pandas as pd
import numpy as np
import os
import datetime as dt

path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/BT results v1/"

files  = os.listdir(path)
files

for y in [2025]:

    for file in files:

        
        try:
            new_path = path + file + '/Final Summary/' + str(y) + "/"
            df = pd.read_excel(new_path + "Best combinations BT summary.xlsx")
            
            df_FT = pd.read_excel(new_path + " Test data summary.xlsx")
            
            df = df.merge(df_FT, how = 'left', on= ['Strategy','Currency','Condition','Timeframe',
              
                                                    'parameter_1','parameter_2','parameter_3','parameter_4',
                                                    'parameter_5','parameter_6'])
            
            df.to_excel(new_path + "BT_FT summary.xlsx",index = False)
            
        except:
            print("Error in", file)
#             break
            continue
