==> S4 - python script to extract all comb summary.ipynb <==
import pandas as pd
import numpy as np
import os

year_list = [2025]

initial_path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/BT results v1/"

exp_path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/All comb summary/"

files = os.listdir(initial_path)

# files = ['123_pattern',
#  '2xIBS RSI 3 Day Strategy',
#  '3 Day Low Overnight',
#  '5 Day Low Overnight',
#  '9_30_strategy',
#  'BB 3 day Strategy',
#  'Big 3 Trading Strategy 1',
#  'Big 3 Trading Strategy 2',
#  'BioTech Strategy',
#  'Bollinger_Band_Squeeze',
#  'Bottom Range - IBS',
#  'Breakouts_3',
#  'Bulls_strategy',
#  'Buy_the_dip',
#  'Cardano',
#  'DDC Mean Reversion',
#  'DDC Momemtum',
#  'n_EMA',
#  'OBV RSI 3 Day Strategy',
#  'Overnight Options Week RSI Edge',
#  'Overnight RSI Edge',
#  'overnight strategy',
#  'Overnight_s1_s2',
#  'Overnight_s3',
#  'Parabolic Sar2 s2',
#  'Reversal_Day',
#  'RSI IBS 3 day Strategy',
#  'Sell the RIP Strategy',
#  'Stoch 3 day Strategy',
#  'Sugar_Futures_strategy']
files


for y in year_list:
    
    all_comb = pd.DataFrame()
    
    for file in files:
        
        try:

            new_path = initial_path + file + "/Final Summary/" + str(y) +"/"

            df = pd.read_excel(new_path + "BT_FT summary.xlsx")

            # all_comb = all_comb.append(df,ignore_index = True)
            all_comb = pd.concat([all_comb,df],ignore_index = True)
        
        except Exception as e:
            print(e)
            continue

    all_comb['Accuracy_BT']  = all_comb['Profitable_Trades_x']/all_comb['Total_Trades_x']
    all_comb['PPT_BT'] =  all_comb['Net_P_L_x']/all_comb['Total_Trades_x']

    all_comb = all_comb.rename(columns = {'Accuracy':'Accuracy_FT','PPT':'PPT_FT'})

    
    
    all_comb.to_excel(exp_path + str(y) + " All Stg BT FT summary.xlsx",index = False)

    
    

all_comb.columns.to_list()
