==> S5 - Python script to filter to summarize BT and FT trades.ipynb <==
import pandas as pd
import numpy as np
import os

path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/All comb summary/"

exp_path = "E:/WSG Markets/Backtesting V2 Nov 2025/Out of sample/All comb summary/Final summary/"

files = os.listdir(path)
files.remove("Final summary")
# files.remove("~$2018 All Stg BT FT summary.xlsx")
# files.remove("~$2019 All Stg BT FT summary.xlsx")
files

all_comb_summary = pd.DataFrame()

for y in [2025]:
    
    all_df = pd.DataFrame()
    
    for file in files:
        
        file.split(" ")[0]

        df = pd.read_excel(path + file)
#         print(len(df), file)
        
        if str(file.split(" ")[0]) == str(y):
            
            all_df = pd.concat([all_df,df],ignore_index = True)
            
#             print(len(all_df), file)
            
            
        
#     all_df = all_df[all_df['Currency'] != 'EXP/USD'].reset_index(drop = True)
#     all_df = all_df[all_df['Currency'] != 'INT/USD'].reset_index(drop = True)
    
    
#     all_df.to_csv(exp_path + "All pair summary " + str(y) + ".csv",index = False)
    print(len(all_df),y)
    
#     all_df = all_df[all_df['Accuracy_x'] >= 0.65].reset_index(drop = True)
            
        
    all_df_1 = pd.DataFrame(all_df.groupby(by = ['Timeframe']).agg(
                                                Net_P_L_BT=('Net_P_L_x', 'sum'),
                                                Net_P_L_FT=('Net_P_L_y', 'sum'),
                                                Total_Trades_BT=('Total_Trades_x', 'sum'),
                                                Total_Trades_FT=('Total_Trades_y', 'sum'),
                                                Profitable_Trades_BT = ('Profitable_Trades_x','sum'),
                                                Profitable_Trades_FT = ('Profitable_Trades_y','sum'))).reset_index()
    
    all_df_1['Year'] = y
    
    all_comb_summary = pd.concat([all_comb_summary,all_df_1],ignore_index = True)
            
        
        


    

all_comb_summary.to_csv(exp_path + "All comb summary across years v1.csv",index = False)
